Move handleInitializeApp to handlers/app/ folder
